refactor: consolidate and refactor utils for handling asset caching based on MD5 hash differences
